+++ src/rgb.h
@@ -2,6 +2,6 @@
 #define RGB_H
 
 void rgb_init(uint16_t pin);
-void rgb_set_color(int red, int green, int blue);
+void rgb_set_color(uint8_t red, uint8_t green, uint8_t blue);
 
 #endif
